Use the failure crate to propagate Tokenizer errors
As a result of this change, invalid tokens will be caught by the
Tokenizer and produce an appropriate error message (such as "Error: `%`
is not a valid token"). Previously, the Parser would simply stop as
there was no difference between an invalid token and a premature EOF.

As part of this work, it was necessary to tweak the Parser so that it
would accept a vector of Tokens, rather than the Tokenizer itself.