Add --init and --no-init flags to configure the program's entry point
By default, the VM translator generates a Hack program that will jump to
the Sys.init function after some simple initialisation steps such as
setting the stack pointer.

The --init flag allows the user to supply a different function name, in
the case of a single-file program or whenever the Sys module isn't being
used.

Conversely, the --no-init flag stops the VM translator jumping to an
entry point at all: instead, execution starts at the top of the input
file. This is only ever likely to be useful when compiling a very simple
.vm file that isn't broken up into multiple functions.